Vector_Method.ipynb
In this jupyter notebook, a method for detecting and parsing out major temperature inversions of interest within an upper air sounding is presented.

First, upper air data is ingested from the University of Wyoming's Atmospheric Soundings website and a Skew-T Log-P chart is built from scratch.

Next, data is scanned from the surface upward until the first point at which the temperature increases with height is detected. This point is saved as the starting point of the temperature inversion. Upward scanning continues until a point in which the temperature returns to decreasing with height is detected. The point preceding this point is saved as the ending point of the temperature inversion. A line is drawn between these two points and labeled as an "inversion vector." Scanning continues until all inversion vectors are grabbed.

Lastly, the temperature and height axes of the Skew-T Log-P chart are scaled to a similar range. The distance of the inversion vectors in the scaled Temperature-Pressure space is ranked and the inversion vector with the largest distance is parsed out and labelled as the primary inversion of interest. In a subsequent notebook, the primary inversions between the High Resolution Rapid Refresh (HRRR) Numerical Weather Prediction (NWP) model and observed data are compared. This notbook will simply present a method for parsing out the primary inversion.